+++ docassemble_base/docassemble/base/util.py
@@ -34,7 +34,7 @@ import shutil
 import subprocess
 from bs4 import BeautifulSoup
 
-__all__ = ['alpha', 'roman', 'item_label', 'ordinal', 'ordinal_number', 'comma_list', 'word', 'get_language', 'set_language', 'get_dialect', 'set_country', 'get_country', 'get_locale', 'set_locale', 'comma_and_list', 'need', 'nice_number', 'quantity_noun', 'currency_symbol', 'verb_past', 'verb_present', 'noun_plural', 'noun_singular', 'indefinite_article', 'capitalize', 'space_to_underscore', 'force_ask', 'force_gather', 'period_list', 'name_suffix', 'currency', 'static_image', 'title_case', 'url_of', 'process_action', 'url_action', 'get_info', 'set_info', 'get_config', 'prevent_going_back', 'qr_code', 'action_menu_item', 'from_b64_json', 'defined', 'value', 'message', 'response', 'json_response', 'command', 'single_paragraph', 'quote_paragraphs', 'location_returned', 'location_known', 'user_lat_lon', 'interview_url', 'interview_url_action', 'interview_url_as_qr', 'interview_url_action_as_qr', 'LatitudeLongitude', 'RoleChangeTracker', 'Name', 'IndividualName', 'Address', 'City', 'Event', 'Person', 'Thing', 'Individual', 'ChildList', 'FinancialList', 'PeriodicFinancialList', 'Income', 'Asset', 'Expense', 'Value', 'PeriodicValue', 'OfficeList', 'Organization', 'objects_from_file', 'send_email', 'send_sms', 'map_of', 'selections', 'DAObject', 'DAList', 'DADict', 'DASet', 'DAFile', 'DAFileCollection', 'DAFileList', 'DAEmail', 'DAEmailRecipient', 'DAEmailRecipientList', 'DATemplate', 'last_access_time', 'last_access_delta', 'last_access_days', 'last_access_hours', 'last_access_minutes', 'action_arguments', 'action_argument', 'timezone_list', 'as_datetime', 'current_datetime', 'date_difference', 'date_interval', 'year_of', 'month_of', 'day_of', 'dow_of', 'format_date', 'format_time', 'today', 'get_default_timezone', 'user_logged_in', 'interface', 'user_privileges', 'user_has_privilege', 'user_info', 'task_performed', 'task_not_yet_performed', 'mark_task_as_performed', 'times_task_performed', 'set_task_counter', 'background_action', 'background_response', 'background_response_action', 'us', 'DARedis', 'MachineLearningEntry', 'SimpleTextMachineLearner', 'SVMMachineLearner', 'set_live_help_status', 'chat_partners_available', 'phone_number_in_e164', 'phone_number_is_valid', 'countries_list', 'country_name', 'write_record', 'read_records', 'delete_record', 'variables_as_json', 'all_variables', 'ocr_file', 'ocr_file_in_background', 'get_sms_session', 'initiate_sms_session', 'terminate_sms_session', 'language_from_browser', 'device', 'interview_email', 'get_emails', 'plain', 'bold', 'italic', 'path_and_mimetype', 'states_list', 'state_name', 'subdivision_type', 'indent', 'raw', 'fix_punctuation']
+__all__ = ['alpha', 'roman', 'item_label', 'ordinal', 'ordinal_number', 'comma_list', 'word', 'get_language', 'set_language', 'get_dialect', 'set_country', 'get_country', 'get_locale', 'set_locale', 'comma_and_list', 'need', 'nice_number', 'quantity_noun', 'currency_symbol', 'verb_past', 'verb_present', 'noun_plural', 'noun_singular', 'indefinite_article', 'capitalize', 'space_to_underscore', 'force_ask', 'force_gather', 'period_list', 'name_suffix', 'currency', 'static_image', 'title_case', 'url_of', 'process_action', 'url_action', 'get_info', 'set_info', 'get_config', 'prevent_going_back', 'qr_code', 'action_menu_item', 'from_b64_json', 'defined', 'value', 'message', 'response', 'json_response', 'command', 'single_paragraph', 'quote_paragraphs', 'location_returned', 'location_known', 'user_lat_lon', 'interview_url', 'interview_url_action', 'interview_url_as_qr', 'interview_url_action_as_qr', 'LatitudeLongitude', 'RoleChangeTracker', 'Name', 'IndividualName', 'IndividualOrBusinessName',  'Address', 'City', 'Event', 'Person', 'Thing', 'Individual', 'IndividualOrBusiness', 'ChildList', 'FinancialList', 'PeriodicFinancialList', 'Income', 'Asset', 'Expense', 'Value', 'PeriodicValue', 'OfficeList', 'Organization', 'objects_from_file', 'send_email', 'send_sms', 'map_of', 'selections', 'DAObject', 'DAList', 'DADict', 'DASet', 'DAFile', 'DAFileCollection', 'DAFileList', 'DAEmail', 'DAEmailRecipient', 'DAEmailRecipientList', 'DATemplate', 'last_access_time', 'last_access_delta', 'last_access_days', 'last_access_hours', 'last_access_minutes', 'action_arguments', 'action_argument', 'timezone_list', 'as_datetime', 'current_datetime', 'date_difference', 'date_interval', 'year_of', 'month_of', 'day_of', 'dow_of', 'format_date', 'format_time', 'today', 'get_default_timezone', 'user_logged_in', 'interface', 'user_privileges', 'user_has_privilege', 'user_info', 'task_performed', 'task_not_yet_performed', 'mark_task_as_performed', 'times_task_performed', 'set_task_counter', 'background_action', 'background_response', 'background_response_action', 'us', 'DARedis', 'MachineLearningEntry', 'SimpleTextMachineLearner', 'SVMMachineLearner', 'set_live_help_status', 'chat_partners_available', 'phone_number_in_e164', 'phone_number_is_valid', 'countries_list', 'country_name', 'write_record', 'read_records', 'delete_record', 'variables_as_json', 'all_variables', 'ocr_file', 'ocr_file_in_background', 'read_qr', 'get_sms_session', 'initiate_sms_session', 'terminate_sms_session', 'language_from_browser', 'device', 'interview_email', 'get_emails', 'plain', 'bold', 'italic', 'path_and_mimetype', 'states_list', 'state_name', 'subdivision_type', 'indent', 'raw', 'fix_punctuation']
 
 #knn_machine_learner = DummyObject
 
@@ -422,6 +422,32 @@ class Name(DAObject):
 #    def __repr__(self):
 #        return(repr(self.full()))
 
+class IndividualOrBusinessName(Name):
+    """The name of an IndividualOrBusiness."""
+    def init(self, *pargs, **kwargs):
+        return super(IndividualOrBusinessName, self).init(*pargs, **kwargs)
+    def full(self, middle='initial', use_suffix=True):
+        """Returns the full name.  Has optional keyword arguments middle 
+        and use_suffix."""
+        if self.is_individual:
+            self.uses_parts = True
+        return super(IndividualOrBusinessName, self).full()
+        # if super(IndividualOrBusinessName, self).defined():
+        #     return super(IndividualOrBusinessName, self).full()
+        # else:
+        #     return self.instanceName
+    def firstlast(self):
+        """Returns the first name followed by the last name."""
+        if self.is_individual:
+            self.uses_parts = True
+        return super(IndividualOrBusinessName, self).firstlast()
+    def lastfirst(self):
+        """Returns the last name followed by a comma, followed by the 
+        last name, followed by the suffix (if a suffix exists)."""
+        if self.is_individual:
+            self.uses_parts = True
+        return super(IndividualOrBusinessName, self).lastfirst()    
+
 class IndividualName(Name):
     """The name of an Individual."""
     def init(self, *pargs, **kwargs):
@@ -787,15 +813,15 @@ class Person(DAObject):
 class Individual(Person):
     """Represents a natural person."""
     def init(self, *pargs, **kwargs):
-        if 'name' not in kwargs:
+        if 'name' not in kwargs and not hasattr(self, 'name'):
             self.name = IndividualName()
-        if 'child' not in kwargs:
+        if 'child' not in kwargs and not hasattr(self, 'child'):
             self.child = ChildList()
-        if 'income' not in kwargs:
+        if 'income' not in kwargs and not hasattr(self, 'income'):
             self.income = Income()
-        if 'asset' not in kwargs:
+        if 'asset' not in kwargs and not hasattr(self, 'asset'):
             self.asset = Asset()
-        if 'expense' not in kwargs:
+        if 'expense' not in kwargs and not hasattr(self, 'expense'):
             self.expense = Expense()
         return super(Individual, self).init(*pargs, **kwargs)
     def identified(self):
@@ -890,6 +916,13 @@ class Individual(Person):
         else:
             return(output)
 
+class IndividualOrBusiness(Individual):
+    """Represents a person that could be an individual or a person."""
+    def init(self, *pargs, **kwargs):
+        if 'name' not in kwargs:
+            self.name = IndividualOrBusinessName()
+        return super(IndividualOrBusiness, self).init(*pargs, **kwargs)
+        
 class ChildList(DAList):
     """Represents a list of children."""
     def init(self, *pargs, **kwargs):
@@ -1401,6 +1434,50 @@ def ocr_file(image_file, language=None, psm=6, f=None, l=None, x=None, y=None, W
     for directory in temp_directory_list:
         shutil.rmtree(directory)
     return "\f".join(page_text)
+
+def read_qr(image_file, f=None, l=None, x=None, y=None, W=None, H=None):
+    """Reads QR codes from a file or files and returns a list of codes found."""
+    import qrtools
+    if not (isinstance(image_file, DAFile) or isinstance(image_file, DAFileList)):
+        return word("(Not a DAFile or DAFileList object)")
+    if isinstance(image_file, DAFile):
+        image_file = [image_file]
+    file_list = list()
+    for doc in image_file:
+        if hasattr(doc, 'extension'):
+            if doc.extension not in ['pdf', 'png', 'jpg', 'gif']:
+                return word("(Not a readable image file)")
+            path = doc.path()
+            if doc.extension == 'pdf':
+                directory = tempfile.mkdtemp()
+                temp_directory_list.append(directory)
+                prefix = os.path.join(directory, 'page')
+                args = [pdf_to_ppm, '-r', str(ocr_resolution)]
+                if f is not None:
+                    args.extend(['-f', str(f)])
+                if l is not None:
+                    args.extend(['-l', str(l)])
+                if x is not None:
+                    args.extend(['-x', str(x)])
+                if y is not None:
+                    args.extend(['-y', str(y)])
+                if W is not None:
+                    args.extend(['-W', str(W)])
+                if H is not None:
+                    args.extend(['-H', str(H)])
+                args.extend(['-png', path, prefix])
+                result = subprocess.call(args)
+                if result > 0:
+                    return word("(Unable to extract images from PDF file)")
+                file_list.extend(sorted([os.path.join(directory, f) for f in os.listdir(directory) if os.path.isfile(os.path.join(directory, f))]))
+                continue
+            file_list.append(path)
+    codes = list()
+    for page in file_list:
+        qr = qrtools.QR()
+        qr.decode(page)
+        codes.append(qr.data)
+    return codes        
 
 def path_and_mimetype(file_ref):
     """Returns a path and the MIME type of a file"""